+++ src/account/HISTORY.rst
@@ -5,9 +5,10 @@ Release History
 
 0.2.4
 +++++
-* Update SDK `azure-mgmt-subscription` version to 3.1.1
+* Update SDK `azure-mgmt-subscription` version to 3.1.1 and api-version to 2021-10-01.
 * GA of `az account alias`: Manage subscription alias.
 * Add new command `az account accept-ownership-status` to support accept subscription ownership status
+* Update `minCliCoreVersion` requirement from `2.3.1` to `2.38.0`.
 
 0.2.1
 +++++