+++ src/cosmosdb-preview/HISTORY.rst
@@ -2,6 +2,9 @@
 
 Release History
 ===============
+0.18.1
+++++++
+* Modify parameter for Managed Identity name.
 
 0.18.0
 ++++++